Refactor mixins into a package.
