+++ pytest_celery/test_services/mixins/__init__.py
@@ -1,0 +1,4 @@
+from pytest_celery.test_services.mixins.rabbitmq import RabbitMQTestServiceMixin
+from pytest_celery.test_services.mixins.redis import RedisTestServiceMixin
+
+__all__ = ("RedisTestServiceMixin", "RabbitMQTestServiceMixin")
